Split metrics and loss functions off of utils
