+++ chemprop/spectra_utils.py
@@ -6,159 +6,6 @@ import torch
 import numpy as np
 
 
-def sid_loss(model_spectra: torch.tensor, target_spectra: torch.tensor, mask: torch.tensor, threshold: float = None) -> torch.tensor:
-    """
-    Loss function for use with spectra data type.
-
-    :param model_spectra: The predicted spectra output from a model with shape (batch_size,spectrum_length).
-    :param target_spectra: The target spectra with shape (batch_size,spectrum_length). Values must be normalized so that each spectrum sums to 1.
-    :param mask: Tensor with boolian indications of where the spectrum output should not be excluded with shape (batch_size,spectrum_length).
-    :param threshold: Loss function requires that values are positive and nonzero. Values below the threshold will be replaced with the threshold value.
-    :return: A tensor containing loss values for the batch with shape (batch_size,spectrum_length).
-    """
-    # Move new tensors to torch device
-    torch_device = model_spectra.device
-
-    # Normalize the model spectra before comparison
-    zero_sub = torch.zeros_like(model_spectra, device=torch_device)
-    one_sub = torch.ones_like(model_spectra, device=torch_device)
-    if threshold is not None:
-        threshold_sub = torch.full(model_spectra.shape,threshold, device=torch_device)
-        model_spectra = torch.where(model_spectra < threshold, threshold_sub, model_spectra)
-    model_spectra = torch.where(mask, model_spectra, zero_sub)
-    sum_model_spectra = torch.sum(model_spectra, axis=1, keepdim=True)
-    model_spectra = torch.div(model_spectra, sum_model_spectra)
-
-    # Calculate loss value
-    target_spectra = torch.where(mask, target_spectra, one_sub)
-    model_spectra = torch.where(mask, model_spectra, one_sub) # losses in excluded regions will be zero because log(1/1) = 0. 
-    loss = torch.mul(torch.log(torch.div(model_spectra, target_spectra)), model_spectra) \
-        + torch.mul(torch.log(torch.div(target_spectra, model_spectra)), target_spectra)
-    loss = loss.to(torch_device)
-
-    return loss
-
-
-def sid_metric(model_spectra: List[List[float]], target_spectra: List[List[float]], threshold: float = None, batch_size: int = 50) -> float:
-    """
-    Metric function for use with spectra data type.
-
-    :param model_spectra: The predicted spectra output from a model with shape (num_data, spectrum_length).
-    :param target_spectra: The target spectra with shape (num_data, spectrum_length). Values must be normalized so that each spectrum sums to 1.
-        Excluded values in target spectra will have a value of None.
-    :param threshold: Function requires that values are positive and nonzero. Values below the threshold will be replaced with the threshold value.
-    :param batch_size: Batch size for calculating metric.
-    :return: The average SID value for the predicted spectra.
-    """
-    losses = []
-    num_iters, iter_step = len(model_spectra), batch_size
-
-    for i in trange(0, num_iters, iter_step):
-
-        # Create batches
-        batch_preds = model_spectra[i:i + iter_step]
-        batch_preds = np.array(batch_preds)
-        batch_targets = target_spectra[i:i + iter_step]
-        batch_mask = np.array([[x is not None for x in b] for b in batch_targets])
-        batch_targets = np.array([[1 if x is None else x for x in b] for b in batch_targets])
-
-        # Normalize the model spectra before comparison
-        if threshold is not None:
-            batch_preds[batch_preds < threshold] = threshold
-        batch_preds[~batch_mask] = 0
-        sum_preds = np.sum(batch_preds, axis=1, keepdims=True)
-        batch_preds = batch_preds / sum_preds
-
-        # Calculate loss value
-        batch_preds[~batch_mask] = 1 # losses in excluded regions will be zero because log(1/1) = 0.
-        loss = batch_preds * np.log(batch_preds / batch_targets) + batch_targets * np.log(batch_targets / batch_preds)
-        loss = np.sum(loss, axis=1)
-
-        # Gather batches
-        loss = loss.tolist()
-        losses.extend(loss)
-
-    loss = np.mean(loss)
-
-    return loss
-
-
-def wasserstein_loss(model_spectra: torch.tensor, target_spectra: torch.tensor, mask: torch.tensor, threshold: float = None) -> torch.tensor:
-    """
-    Loss function for use with spectra data type. This loss assumes that values are evenly spaced.
-
-    :param model_spectra: The predicted spectra output from a model with shape (batch_size,spectrum_length).
-    :param target_spectra: The target spectra with shape (batch_size,spectrum_length). Values must be normalized so that each spectrum sums to 1.
-    :param mask: Tensor with boolian indications of where the spectrum output should not be excluded with shape (batch_size,spectrum_length).
-    :param threshold: Loss function requires that values are positive and nonzero. Values below the threshold will be replaced with the threshold value.
-    :return: A tensor containing loss values for the batch with shape (batch_size,spectrum_length).
-    """
-    # Move new tensors to torch device
-    torch_device = model_spectra.device
-
-    # Normalize the model spectra before comparison
-    zero_sub = torch.zeros_like(model_spectra, device=torch_device)
-    if threshold is not None:
-        threshold_sub = torch.full(model_spectra.shape,threshold, device=torch_device)
-        model_spectra = torch.where(model_spectra < threshold, threshold_sub, model_spectra)
-    model_spectra = torch.where(mask, model_spectra, zero_sub)
-    sum_model_spectra = torch.sum(model_spectra, axis=1, keepdim=True)
-    model_spectra = torch.div(model_spectra, sum_model_spectra)
-
-    # Calculate loss value
-    target_cum = torch.cumsum(target_spectra,axis=1).to(torch_device)
-    model_cum = torch.cumsum(model_spectra,axis=1).to(torch_device)
-    loss = torch.abs(target_cum - model_cum)
-    loss = loss.to(torch_device)
-
-    return loss
-
-
-def wasserstein_metric(model_spectra: List[List[float]], target_spectra: List[List[float]], threshold: float = None, batch_size: int = 50) -> float:
-    """
-    Metric function for use with spectra data type. This metric assumes that values are evenly spaced.
-
-    :param model_spectra: The predicted spectra output from a model with shape (num_data, spectrum_length).
-    :param target_spectra: The target spectra with shape (num_data, spectrum_length). Values must be normalized so that each spectrum sums to 1.
-        Excluded values in target spectra will have value None.
-    :param threshold: Function requires that values are positive and nonzero. Values below the threshold will be replaced with the threshold value.
-    :param batch_size: Batch size for calculating metric.
-    :return: The average wasserstein loss value for the predicted spectra.
-    """
-    losses = []
-    num_iters, iter_step = len(model_spectra), batch_size
-
-    for i in trange(0, num_iters, iter_step):
-
-        # Create batches
-        batch_preds = model_spectra[i:i + iter_step]
-        batch_preds = np.array(batch_preds)
-        batch_targets = target_spectra[i:i + iter_step]
-        batch_mask = np.array([[x is not None for x in b] for b in batch_targets])
-        batch_targets = np.array([[0 if x is None else x for x in b] for b in batch_targets])
-
-        # Normalize the model spectra before comparison
-        if threshold is not None:
-            batch_preds[batch_preds < threshold] = threshold
-        batch_preds[~batch_mask] = 0
-        sum_preds = np.sum(batch_preds, axis=1, keepdims=True)
-        batch_preds = batch_preds / sum_preds
-
-        # Calculate loss value
-        target_cum = np.cumsum(batch_targets,axis=1)
-        preds_cum = np.cumsum(batch_preds,axis=1)
-        loss = np.abs(target_cum - preds_cum)
-        loss = np.sum(loss, axis=1)
-
-        # Gather batches
-        loss = loss.tolist()
-        losses.extend(loss)
-
-    loss = np.mean(loss)
-
-    return loss
-
-
 def normalize_spectra(spectra: List[List[float]], phase_features: List[List[float]] = None, phase_mask: List[List[float]] = None, batch_size: int = 50, excluded_sub_value: float = None, threshold: float = None) -> List[List[float]]:
     """
     Function takes in spectra and normalize them to sum values to 1. If provided with phase mask information, will remove excluded spectrum regions.
