+++ chemprop/utils.py
@@ -3,16 +3,13 @@ import csv
 from datetime import timedelta
 from functools import wraps
 import logging
-import math
 import os
 import pickle
 import re
 from time import time
-from typing import Any, Callable, List, Tuple, Union
+from typing import Any, Callable, List, Tuple
 import collections
 
-from sklearn.metrics import auc, mean_absolute_error, mean_squared_error, precision_recall_curve, r2_score,\
-    roc_auc_score, accuracy_score, log_loss
 import torch
 import torch.nn as nn
 from torch.optim import Adam, Optimizer
@@ -23,7 +20,6 @@ from chemprop.args import PredictArgs, TrainArgs
 from chemprop.data import StandardScaler, MoleculeDataset, preprocess_smiles_columns, get_task_names
 from chemprop.models import MoleculeModel
 from chemprop.nn_utils import NoamLR
-from chemprop.spectra_utils import sid_loss, sid_metric, wasserstein_loss, wasserstein_metric
 
 
 def makedirs(path: str, isfile: bool = False) -> None:
@@ -265,7 +261,8 @@ def load_frzn_model(model: torch.nn,
     model.load_state_dict(model_state_dict)
     
     return model
-    
+
+
 def load_scalers(path: str) -> Tuple[StandardScaler, StandardScaler, StandardScaler, StandardScaler]:
     """
     Loads the scalers a model was trained with.
@@ -320,158 +317,6 @@ def load_task_names(path: str) -> List[str]:
     :return: A list of the task names that the model was trained with.
     """
     return load_args(path).task_names
-
-
-def get_loss_func(args: TrainArgs) -> nn.Module:
-    """
-    Gets the loss function corresponding to a given dataset type.
-
-    :param args: Arguments containing the dataset type ("classification", "regression", or "multiclass").
-    :return: A PyTorch loss function.
-    """
-    if args.alternative_loss_function is not None:
-        if args.dataset_type == 'spectra' and args.alternative_loss_function == 'wasserstein':
-            return wasserstein_loss
-        else:
-            raise ValueError(f'Alternative loss function {args.alternative_loss_function} not '
-                                'supported with dataset type {args.dataset_type}.')
-
-    if args.dataset_type == 'classification':
-        return nn.BCEWithLogitsLoss(reduction='none')
-
-    if args.dataset_type == 'regression':
-        return nn.MSELoss(reduction='none')
-
-    if args.dataset_type == 'multiclass':
-        return nn.CrossEntropyLoss(reduction='none')
-
-    if args.dataset_type == 'spectra':
-        return sid_loss
-
-    raise ValueError(f'Dataset type "{args.dataset_type}" not supported.')
-
-
-def prc_auc(targets: List[int], preds: List[float]) -> float:
-    """
-    Computes the area under the precision-recall curve.
-
-    :param targets: A list of binary targets.
-    :param preds: A list of prediction probabilities.
-    :return: The computed prc-auc.
-    """
-    precision, recall, _ = precision_recall_curve(targets, preds)
-    return auc(recall, precision)
-
-
-def bce(targets: List[int], preds: List[float]) -> float:
-    """
-    Computes the binary cross entropy loss.
-
-    :param targets: A list of binary targets.
-    :param preds: A list of prediction probabilities.
-    :return: The computed binary cross entropy.
-    """
-    # Don't use logits because the sigmoid is added in all places except training itself
-    bce_func = nn.BCELoss(reduction='mean')
-    loss = bce_func(target=torch.Tensor(targets), input=torch.Tensor(preds)).item()
-
-    return loss
-
-
-def rmse(targets: List[float], preds: List[float]) -> float:
-    """
-    Computes the root mean squared error.
-
-    :param targets: A list of targets.
-    :param preds: A list of predictions.
-    :return: The computed rmse.
-    """
-    return math.sqrt(mean_squared_error(targets, preds))
-
-
-def mse(targets: List[float], preds: List[float]) -> float:
-    """
-    Computes the mean squared error.
-
-    :param targets: A list of targets.
-    :param preds: A list of predictions.
-    :return: The computed mse.
-    """
-    return mean_squared_error(targets, preds)
-
-
-def accuracy(targets: List[int], preds: Union[List[float], List[List[float]]], threshold: float = 0.5) -> float:
-    """
-    Computes the accuracy of a binary prediction task using a given threshold for generating hard predictions.
-
-    Alternatively, computes accuracy for a multiclass prediction task by picking the largest probability.
-
-    :param targets: A list of binary targets.
-    :param preds: A list of prediction probabilities.
-    :param threshold: The threshold above which a prediction is a 1 and below which (inclusive) a prediction is a 0.
-    :return: The computed accuracy.
-    """
-    if type(preds[0]) == list:  # multiclass
-        hard_preds = [p.index(max(p)) for p in preds]
-    else:
-        hard_preds = [1 if p > threshold else 0 for p in preds]  # binary prediction
-
-    return accuracy_score(targets, hard_preds)
-
-
-def get_metric_func(metric: str) -> Callable[[Union[List[int], List[float]], List[float]], float]:
-    r"""
-    Gets the metric function corresponding to a given metric name.
-
-    Supports:
-
-    * :code:`auc`: Area under the receiver operating characteristic curve
-    * :code:`prc-auc`: Area under the precision recall curve
-    * :code:`rmse`: Root mean squared error
-    * :code:`mse`: Mean squared error
-    * :code:`mae`: Mean absolute error
-    * :code:`r2`: Coefficient of determination R\ :superscript:`2`
-    * :code:`accuracy`: Accuracy (using a threshold to binarize predictions)
-    * :code:`cross_entropy`: Cross entropy
-    * :code:`binary_cross_entropy`: Binary cross entropy
-
-    :param metric: Metric name.
-    :return: A metric function which takes as arguments a list of targets and a list of predictions and returns.
-    """
-    if metric == 'auc':
-        return roc_auc_score
-
-    if metric == 'prc-auc':
-        return prc_auc
-
-    if metric == 'rmse':
-        return rmse
-
-    if metric == 'mse':
-        return mse
-
-    if metric == 'mae':
-        return mean_absolute_error
-
-    if metric == 'r2':
-        return r2_score
-
-    if metric == 'accuracy':
-        return accuracy
-
-    if metric == 'cross_entropy':
-        return log_loss
-
-    if metric == 'binary_cross_entropy':
-        return bce
-    
-    if metric == 'sid':
-        return sid_metric
-    
-    if metric == 'wasserstein':
-        return wasserstein_metric
-
-    raise ValueError(f'Metric "{metric}" not supported.')
 
 
 def build_optimizer(model: nn.Module, args: TrainArgs) -> Optimizer:
